[Refactor] Refactor Java UDF context glue
