+++ be/src/exprs/agg/java_udaf_function.h
@@ -34,6 +34,7 @@
 #include "types/logical_type.h"
 #include "udf/java/java_data_converter.h"
 #include "udf/java/java_udf.h"
+#include "udf/java/java_udf_context.h"
 #include "udf/java/java_udf_reflection.h"
 
 namespace starrocks {
